refactor: migrate signal support modules into drift-engine
(cherry picked from commit d85f2cbfcca5b023386763013715bb8c0e5033db)
