+++ src/drift/response_shaping.py
@@ -1,232 +1,9 @@
-"""Response envelope, schema version, and profile-based field filtering.
+"""Re-export stub -- canonical implementation lives in drift_engine.response_shaping."""
+# ruff: noqa: F401, F403
 
-Provides the structural framing for all Drift API responses — schema
-versioning, scope descriptors, and response-profile filtering (ADR-025
-Phase B).
-"""
+import importlib as _importlib
+import sys as _sys
 
-from __future__ import annotations
+from drift_engine.response_shaping import *
 
-from typing import Any
-
-SCHEMA_VERSION = "2.2"
-
-
-def build_drift_score_scope(
-    *,
-    context: str,
-    path: str | None = None,
-    signal_scope: str = "all",
-    baseline_filtered: bool = False,
-) -> str:
-    """Return a stable scope descriptor for drift_score values."""
-    normalized_path = (path or "all").strip("/") or "all"
-    parts = [
-        f"context:{context}",
-        f"signals:{signal_scope}",
-        f"path:{normalized_path}",
-    ]
-    if baseline_filtered:
-        parts.append("baseline:filtered")
-    return ",".join(parts)
-
-
-def _base_response(**extra: Any) -> dict[str, Any]:
-    """Build the common response envelope."""
-    return {"schema_version": SCHEMA_VERSION, **extra}
-
-
-# ---------------------------------------------------------------------------
-# Response profiles (ADR-025 Phase B) — typed handoffs
-# ---------------------------------------------------------------------------
-
-VALID_RESPONSE_PROFILES = ("planner", "coder", "verifier", "merge_readiness")
-
-# Fields that each profile retains.  Unlisted fields are stripped.
-# "schema_version", "status", "agent_instruction" are always kept.
-_PROFILE_KEEP: dict[str, frozenset[str]] = {
-    "planner": frozenset(
-        {
-            "task_graph",
-            "workflow_plan",
-            "tasks",
-            "execution_phases",
-            "next_tool_call",
-            "fallback_tool_call",
-            "done_when",
-            "drift_score",
-            "severity",
-            "finding_count",
-            "top_signals",
-            "scope",
-            "guardrails",
-            "guardrails_prompt_block",
-            "risk_summary",
-            "landscape",
-            "trend",
-            "warnings",
-            "session",
-        }
-    ),
-    "coder": frozenset(
-        {
-            "findings",
-            "finding_count",
-            "fix_first",
-            "tasks",
-            "drift_score",
-            "severity",
-            "negative_context",
-            "items_returned",
-            "next_tool_call",
-            "fallback_tool_call",
-            "done_when",
-            "guardrails",
-            "guardrails_prompt_block",
-            "session",
-        }
-    ),
-    "verifier": frozenset(
-        {
-            "drift_score",
-            "score_delta",
-            "safe_to_commit",
-            "direction",
-            "confidence_map",
-            "severity",
-            "new_findings",
-            "resolved_findings",
-            "accept_change",
-            "blocking_reasons",
-            "done_when",
-            "next_tool_call",
-            "fallback_tool_call",
-            "tasks",
-            "drift_detected",
-            "resolved_count",
-            "new_count",
-            "resolved_count_by_rule",
-            "session",
-        }
-    ),
-    "merge_readiness": frozenset(
-        {
-            "drift_score",
-            "severity",
-            "score_delta",
-            "drift_detected",
-            "accept_change",
-            "blocking_reasons",
-            "warnings",
-            "trend",
-            "finding_count",
-            "top_signals",
-            "done_when",
-            "session",
-        }
-    ),
-}
-
-_ALWAYS_KEEP = frozenset(
-    {
-        "schema_version",
-        "status",
-        "agent_instruction",
-        "response_profile",
-    }
-)
-
-
-def shape_for_profile(
-    result: dict[str, Any],
-    profile: str | None,
-) -> dict[str, Any]:
-    """Filter response fields to match a response profile.
-
-    If *profile* is ``None`` or unrecognised the result is returned unchanged
-    (with a ``response_profile`` key set to ``"full"``).
-    """
-    if not profile or profile not in _PROFILE_KEEP:
-        result["response_profile"] = "full"
-        return result
-
-    keep = _PROFILE_KEEP[profile] | _ALWAYS_KEEP
-    shaped = {k: v for k, v in result.items() if k in keep}
-    shaped["response_profile"] = profile
-    return shaped
-
-
-# ---------------------------------------------------------------------------
-# Output mode: mirror — diagnostic facts only, no prescriptive guidance
-# ---------------------------------------------------------------------------
-
-# Top-level response keys that carry prescriptive guidance (tool choreography,
-# repair instructions, workflow orchestration).  Stripped in mirror mode.
-_PRESCRIPTIVE_TOP_KEYS = frozenset(
-    {
-        "agent_instruction",
-        "next_tool_call",
-        "fallback_tool_call",
-        "done_when",
-        "workflow_plan",
-        "recommended_next_actions",
-        "guardrails",
-        "guardrails_prompt_block",
-        "negative_context",
-    }
-)
-
-# Per-task keys that prescribe how to fix rather than what is wrong.
-_PRESCRIPTIVE_TASK_KEYS = frozenset(
-    {
-        "action",
-        "constraints",
-        "success_criteria",
-        "verify_plan",
-        "expected_effect",
-        "negative_context",
-        "regression_guidance",
-        "repair_exemplar",
-        "fix_intent",
-        "fix_template_class",
-        "repair_maturity",
-    }
-)
-
-
-def apply_output_mode(
-    result: dict[str, Any],
-    mode: str,
-) -> dict[str, Any]:
-    """Apply output-mode filtering to a response dict.
-
-    In ``"full"`` mode (default) the response is returned unchanged.
-
-    In ``"mirror"`` mode all prescriptive guidance is stripped — what
-    remains are diagnostic structural facts: scores, findings, deltas,
-    related files, automation fitness, and severity classifications.
-    The agent receives a structural mirror of the codebase state and
-    decides autonomously what to do with the information.
-    """
-    if mode != "mirror":
-        result.setdefault("output_mode", "full")
-        return result
-
-    result["output_mode"] = "mirror"
-
-    # Strip top-level prescriptive keys
-    for key in _PRESCRIPTIVE_TOP_KEYS:
-        result.pop(key, None)
-
-    # Strip prescriptive fields from task dicts
-    if "tasks" in result and isinstance(result["tasks"], list):
-        for task in result["tasks"]:
-            if isinstance(task, dict):
-                for key in _PRESCRIPTIVE_TASK_KEYS:
-                    task.pop(key, None)
-
-    # Strip nudge-specific prescriptive text
-    if "nudge" in result and isinstance(result["nudge"], str):
-        result.pop("nudge", None)
-
-    return result
+_sys.modules[__name__] = _importlib.import_module("drift_engine.response_shaping")
